Moved test to test/display_test.py
